float seems better for this application

+++ OK.Siemens.Models/DataRecord.cs
@@ -2,5 +2,5 @@
 
 public class DataRecord: DataRecordBase
 {
-    public double Value { get; set; }
+    public float Value { get; set; }
 }
